desafio3-docker-compose/web/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
colorama==0.4.6
